Created pavement scripts for all examples , removed old pavement scripts, and updated  Readme
